requirements.txt
openai==1.3.6 
panel==1.3.4
trafilatura
pyautogen